requirements.txt
python-dotenv
tensorflow
pillow
numpy
streamlit
keras
opencv-python